Further split into classes.
